requirements.txt
streamlit
requests
python-dotenv
pandas
matplotlib
openai
camel-ai
